Moving files and folders to their respective spaces
