--- README.md
@@ -1,6 +1,0 @@
-# app_dev
-Class project for application development
-
-
-=======
-## Application: Space run
