+++ backend/package.json
@@ -1,22 +1,27 @@
 {
-  "name": "backend",
+  "name": "app_dev_project",
   "version": "1.0.0",
-  "main": "index.js",
+  "description": "",
+  "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon server.js",
-    "start": "node server.js"
+    "start": "node server",
+    "server": "nodemon server"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "bcrypt": "^5.1.1",
-    "dotenv": "^16.0.1",
+    "bcryptjs": "^2.4.3",
+    "config": "^3.3.9",
     "express": "^4.18.2",
+    "express-validator": "^7.0.1",
+    "gravatar": "^1.8.2",
     "jsonwebtoken": "^9.0.2",
-    "mongoose": "^6.12.1",
-    "validator": "^13.11.0"
+    "mongoose": "^8.0.0",
+    "request": "^2.88.2"
   },
-  "description": ""
+  "devDependencies": {
+    "concurrently": "^8.2.2",
+    "nodemon": "^3.0.1"
+  }
 }
