--- models/User.js
@@ -1,27 +1,0 @@
-const mongoose = require('mongoose');
-
-const UserSchema = new mongoose.Schema({
-    name: {
-        type: String,
-        required: true
-    },
-    email: {
-        type: String, 
-        required: true,
-        unique: true
-    },
-    password: {
-        type: String,
-        required : true
-    },
-    avatar: {
-        type: String
-    },
-    date: {
-        type: Date,
-        default: Date.now
-    }
-});
-
-// model name, schema
-module.exports = User = mongoose.model('user', UserSchema);
