--- routes/api/posts.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-//@route  GET api/posts
-//@desc   TEST route
-//@access Public
-router.get('/', (req, res) => res.send('Posts route'));
-
-module.exports = router;
