--- routes/api/profile.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-//@route  GET api/profile
-//@desc   TEST route
-//@access Public
-router.get('/', (req, res) => res.send('Profile route'));
-
-module.exports = router;
